Fixed warning messages
git-svn-id: svn+ssh://svn.cern.ch/reps/geant4/trunk/geant4@47454 82ad68bf-922b-4865-b672-c5fc3b8c97f6

+++ source/processes/hadronic/models/lend/src/gString.cc
@@ -53,7 +53,8 @@ int gString_initialize( statusMessageReporting *smr, gString *gStr, int size, in
         gStr->gStr = NULL;
     }
     gStr->allocated = size;
-    if( increment < gString_minIncrement ) increment = gString_minIncrement;
+    //if( increment < gString_minIncrement ) increment = gString_minIncrement;
+    if( increment < (int) gString_minIncrement ) increment = gString_minIncrement;
     gStr->increment = increment;
     return( 0 );
 }
